Only daySummary uses date in future
