+++ apps/skolplattformen-app/utils/calendarHelpers.ts
@@ -2,16 +2,17 @@ import moment from 'moment'
 
 export const getMeaningfulStartingDate = (date = moment()) => {
   const originalDate = date.clone()
+  let returnDate = date.clone()
   // are we on the evening?
-  if (date.hour() > 17) date = date.add('1', 'day')
+  if (date.hour() > 17) returnDate.add('1', 'day')
   // are we on the weekend
-  if (date.isoWeekday() > 5) {
-    date = date.add(5, 'days').startOf('isoWeek')
-    date
+  if (returnDate.isoWeekday() > 5) {
+    returnDate = returnDate.add(5, 'days').startOf('isoWeek')
+    returnDate
       .hour(originalDate.hour())
       .minute(originalDate.minute())
       .second(originalDate.second())
   }
 
-  return date
+  return returnDate
 }
